FIX CSH SEGEMENTATION VIOLATION
I have run into the csh segmentation violation bug, and I was too lazy to
install the simple "alloc + 4" bugfix, so I tracked it down.  It seems to come
about whenever you are in a directory in your account which has a pathname
length which is divisable by 4, for instance,  /users/kml/g.  To test,
create such a directory, enter it, and sorta execute alot of commands.
Sooner or later, csh will die.

The fix is to add one to the space allocated for a string in the
dcanon() routine where it normalizes directory names which include
the string defined for the environment variable 'HOME'.

AUTHOR:		Kevin Lahey (kml@mosquito.cis.ufl.edu)
386BSD-Patchkit: patch00013