Simple fix for a bug in find from Jeffrey Hsu:
I was intrigued by this problem and looked into it.  The problem is that
stdout for find is block-buffered, since the output is redirected to a
pipe.  But all the grep subprocesses that -exec runs write their output
immediately upon exit.  Only when find has processed enough files to print
BUFSIZ characters or when it exits does find finally print out the list of
filenames.  The only solution I can come up with at this time of the night
is the obvious one.

...set the buffering mode to line.