Fix for # comments as first word on line causing stock /etc/phones to be broken.  This is not a complete fix, but it works for now.
>From: Todd Pfaff <todd@flex.eng.mcmaster.ca>
Subject: Re: freefall commit: '386BSD/src/etc phones'

Yup, a closer look reveals that this code will complain of
an unrecognizable host name if there is only one word on a line
of phones, hence a # by itself screws up.

A simple-minded solution is to add:

   if(*string == '#')
       continue;

as the first statement inside the while loop.