>From: batie@agora.rain.com (Alan Batie) Subject: bug in elvis man page
The elvis man page (and "vi -r") says to use "elvrec" to recover a file
that was being edited when you got dumped off the system or it crashed.
The actual command is "elvisrecover", at least as installed on FreeBSD.

Also, there needs to be a man page link to vi.1 so you can say "man vi".
[editors note, link not done, we do have a real vi now...]