Fixes:
If timed is running when system clock is changed by date command,
improper wtmp entry is made.  According to wtmp(5), two entries, one
with "|" as ut_line field and one with "{" for ut_line, should be
recorded, but, one with "|" and one with "}" are made.

Closes: PR#bin/1182

Submitted by:	 Masafumi NAKANE <masafumi@tky007.tth.expo96.ad.jp>