Some cleanups and optimizations: - Use the system headers method for Elf32/Elf64 symbol compatability - get rid of the UPRINTF debugging. - check the ELF header for compatability much more completely - optimize the section mapper.  Use the same direct VM interfaces that   imgact_aout.c and kern_exec.c use. - Check the return codes from the vm_* functions better.  Some return   KERN_* results, not an errno. - prefault the page tables to reduce startup faults on page tables like   a.out does. - reset the segment protection to zero for each loop, otherwise each   segment could get progressively more privs. (eg: if the first was   read/write/execute, and the second was meant to be read/execute, the   bug would make the second r/w/x too.  In practice this was not a   problem because executables are normally laid out with text first.) - Don't impose arbitary limits.  Use the limits on headers imposed by   the need to fit them into one page. - Remove unused switch() cases now that the verbose debugging is gone.
I've been using an earlier version of this for a month or so.
This sped up ELF exec speed a bit for me but I found it hard to get
consistant benchmarks when I tested it last (a few weeks ago).
I'm still bothered by the page read out of order caused by the
transition from data to bss. This which requires either part filling the
transition page or clearing the remainder.